Extract intance logic [WIP]
