+++ packages/env-instance-task-driver/internal/instance/socket.go
@@ -1,4 +1,4 @@
-package env
+package instance
 
 import (
 	"fmt"
